requirements.txt
openai
pinecone-client
langchain
langchain-community
tiktoken
datasets
youtube-transcript-api
gradio
